Test connectivity in environments where docker containers are not localhost
Connectivity is tested at the end of start.sh and in status.sh.
It was previously achieved by checking if a HTTPS request to localhost would return a HTTP code 200
In the case the docker daemon (and this the bridge network) are not on the same machine as the docker client (foe example in Nanocloud's CI environment) this test always failed.
This fixes the connectivity tests by testing the IP of the proxy container