Fix Download VM did not work in dev mode
In dev mode /var/lib/nanocloud/downloads ans /var/lib/nanocloud/images are virtually not on the same partition because images folder is bind to the host filesystem.
Because of this particularity one cannot just move one file to the other directory but instead needs to copy it and then remove this original one on success.